+++ frontend/src/lib/api_calls/cusomter.js
@@ -22,12 +22,3 @@ export async function makeClaim(itemId, quantity) {
 		return false;
 	}
 }
-
-export function formatDate(dateString) {
-	const date = new Date(dateString);
-	return date.toLocaleDateString('en-US', {
-		year: 'numeric',
-		month: 'short',
-		day: 'numeric'
-	});
-}
