+++ frontend/src/lib/helpers.js
@@ -1,0 +1,8 @@
+export function formatDate(dateString) {
+	const date = new Date(dateString);
+	return date.toLocaleDateString('en-US', {
+		year: 'numeric',
+		month: 'short',
+		day: 'numeric'
+	});
+}
