+++ frontend/src/routes/+page.svelte
@@ -1,8 +1,8 @@
 <script>
 	import { onMount } from 'svelte';
 	import { initializeUserStore } from '$lib/stores/auth';
-	import Items from '$lib/root/Items.svelte';
-	import ClaimedItmes from '$lib/root/ClaimedItmes.svelte';
+	import Items from './root_components/Items.svelte';
+	import ClaimedItmes from './root_components/ClaimedItmes.svelte';
 	import { refreshItems } from '$lib/stores/items';
 
 	onMount(() => {
