+++ frontend/src/routes/admin/dashboard/components/Dashboard.svelte
