+++ frontend/src/routes/admin/inventory/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-	import CreateItem from '$lib/admin/inventory/CreateItem.svelte';
-	import ItemsAdmin from '$lib/admin/inventory/ItemsAdmin.svelte';
-	import ClaimedItemsAdmin from '$lib/admin/inventory/ClaimedItemsAdmin.svelte';
+	import CreateItem from './components/CreateItem.svelte';
+	import ItemsAdmin from './components/ItemsAdmin.svelte';
+	import ClaimedItemsAdmin from './components/ClaimedItemsAdmin.svelte';
 </script>
 
 <CreateItem />
