+++ frontend/src/routes/admin/inventory/components/ClaimItemModal.svelte
@@ -3,7 +3,7 @@
 	import { formatDate } from '$lib/root/helpers.js';
 	import { claimItem } from './helpers.js';
 	import { refreshItems } from '$lib/stores/items';
-	import { users, refreshUsers } from '../users/helpers.js';
+	import { users, refreshUsers } from '$lib/stores/users';
 	let { showClaimModal = $bindable(false), selectedItem = null } = $props();
 
 	let amount = $state(0);
