+++ frontend/src/routes/admin/inventory/components/CreateItem.svelte
