+++ frontend/src/routes/admin/inventory/components/EditItemModal.svelte
