+++ frontend/src/routes/admin/users/components/DeleteUser.svelte
