+++ frontend/src/routes/root_components/Items.svelte
