+++ frontend/src/routes/root_components/MakeClaimModal.svelte
