+++ frontend/src/routes/root_components/PhotoModal.svelte
